App/requirements.txt
accelerate==1.5.2
annotated-types==0.7.0
awscli==1.38.11
blinker==1.9.0
blis==1.2.0
boto3==1.37.11
botocore==1.37.11
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
docutils==0.16
filelock==3.18.0
Flask==3.1.0
fsspec==2025.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
huggingface-hub==0.29.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pillow==11.1.0
preshed==3.0.9
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pt_core_news_lg @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.8.0/pt_core_news_lg-3.8.0-py3-none-any.whl#sha256=2561c9a72a938d37141e9694e1a36d25061a44ce7e4f3bad2d3fa3bb836191af
pt_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.8.0/pt_core_news_sm-3.8.0-py3-none-any.whl#sha256=c304fa04db3af73cd08a250feacf560506e15a2ec2469bd1b09f06847f6b455c
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.7.2
s3transfer==0.11.4
safetensors==0.5.3
sentencepiece==0.2.0
setuptools==76.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.13.1
thinc==8.3.4
tokenizers==0.21.1
torch==2.6.0+cu118
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
typer==0.15.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
